+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 ### Added
 
 - Added `obj` setter to context builder as an alternative to `currentContext.obj`
+- Added `option().boolean()` and `argument().boolean()`
 
 ### Changed
 - Updated Kotlin to 1.6.20